--- PADO/PADO/iOS/Helpers/Enums/ButtonType.swift
@@ -1,13 +1,0 @@
-//
-//  ButtonType.swift
-//  PADO
-//
-//  Created by 강치우 on 3/7/24.
-//
-
-import Foundation
-
-enum ButtonType {
-    case direct
-    case unDirect
-}
