--- PADO/PADO/iOS/Helpers/Enums/CommentType.swift
@@ -1,13 +1,0 @@
-//
-//  CommentType.swift
-//  PADO
-//
-//  Created by 강치우 on 3/7/24.
-//
-
-import Foundation
-
-enum CommentType {
-    case comment
-    case replyComment
-}
