--- PADO/PADO/iOS/Helpers/Enums/EmotionType.swift
@@ -1,41 +1,0 @@
-//
-//  EmotionType.swift
-//  PADO
-//
-//  Created by 강치우 on 3/7/24.
-//
-
-import SwiftUI
-
-enum Emotion: String, CaseIterable {
-    case basic = ""
-    case thumbsUp = "👍"
-    case heart = "🥰"
-    case laughing = "🤣"
-    case angry = "😡"
-    case sad = "😢"
-    case overEat = "🤮"
-    
-    var emoji: String {
-        self.rawValue
-    }
-    
-    var color: Color {
-        switch self {
-        case .basic:
-            return .white
-        case .thumbsUp:
-            return .green
-        case .heart:
-            return .pink
-        case .laughing:
-            return .yellow
-        case .angry:
-            return .orange
-        case .sad:
-            return .blue
-        case .overEat:
-            return .purple
-        }
-    }
-}
