--- PADO/PADO/iOS/Helpers/Enums/FollowerModalType.swift
@@ -1,13 +1,0 @@
-//
-//  FollowerModalType.swift
-//  PADO
-//
-//  Created by 강치우 on 3/7/24.
-//
-
-import Foundation
-
-enum FollowerModalType {
-    case surfer
-    case follower
-}
