--- PADO/PADO/iOS/Helpers/Enums/InputPostType.swift
@@ -1,14 +1,0 @@
-//
-//  InputPostType.swift
-//  PADO
-//
-//  Created by 강치우 on 3/7/24.
-//
-
-import Foundation
-
-enum InputPostType {
-    case pado
-    case sendPado
-    case highlight
-}
