--- PADO/PADO/iOS/Helpers/Enums/PostViewType.swift
@@ -1,14 +1,0 @@
-//
-//  PostViewType.swift
-//  PADO
-//
-//  Created by 강치우 on 3/7/24.
-//
-
-import Foundation
-
-enum PostViewType {
-    case receive
-    case send
-    case highlight
-}
