--- PADO/PADO/iOS/Helpers/Enums/SearchFollowType.swift
@@ -1,13 +1,0 @@
-//
-//  SearchFollowType.swift
-//  PADO
-//
-//  Created by 강치우 on 3/7/24.
-//
-
-import Foundation
-
-enum SearchFollowType {
-    case follower
-    case following
-}
