--- PADO/PADO/iOS/Helpers/Enums/SufferSetType.swift
@@ -1,13 +1,0 @@
-//
-//  SufferSetType.swift
-//  PADO
-//
-//  Created by 강치우 on 3/7/24.
-//
-
-import Foundation
-
-enum SufferSet: String {
-    case removesuffer = "서퍼 해제"
-    case setsuffer = "서퍼 등록"
-}
